+++ internal/features/auth/repository/postgres/models.go
@@ -3,8 +3,8 @@ package auth_repository_postgres
 import "time"
 
 type UserModel struct {
-	ID           int
-	Version      int
+	ID           int64
+	Version      int64
 	Email        string
 	PasswordHash string
 	CreatedAt    time.Time
