+++ internal/features/users/repository/postgres/get_user.go
@@ -10,7 +10,7 @@ import (
 	core_postgres_pool "github.com/loundxr/expense-tracker/internal/core/repository/postgres/pool"
 )
 
-func (r *UsersRepository) GetUserByID(ctx context.Context, id int) (domain.User, error) {
+func (r *UsersRepository) GetUserByID(ctx context.Context, id int64) (domain.User, error) {
 	const op = "users.repository.postgres.GetUser"
 	ctx, cancel := context.WithTimeout(ctx, r.pool.OperationTimeout())
 	defer cancel()
